remove sourcemaps and add comment to `behaviorImported`
Why remove sourcemaps? Surely they're important?
The compiled code is not minified and all exports
are named and so we don't really need the sourcemap.

Hopefully. I do know that including the sourcemaps
brings a problem with webpack compiling within an
application if you don't specify the `process.env`
variables for the dynamic loaded behaviors which I 
think is caused by a mismatch between the compressed
build which will skip the dynamic imports if you 
don't set those variables and the sourcemap which
is expecting them.

I want to make dynamic imports optional and keep
backwards compatibility and right now we have this.

Maybe `0.1.10` will have a better solution!

Happy New Year 🎉